♻️ 코드 리뷰 반영
