+++ Alarmi/Alarmi/Sources/Global/Extensions/UITabBarItem+.swift
@@ -10,11 +10,6 @@ import UIKit
 
 extension UITabBarItem {
     
-    convenience init(title: String, imageName: String) {
-        let image = UIImage(systemName: imageName)
-        self.init(title: title, image: image, selectedImage: nil)
-    }
-
     convenience init(item: MainTabBarController.TabItem) {
         let image = UIImage(systemName: item.imageName)
         self.init(title: item.title, image: image, selectedImage: nil)
