+++ Alarmi/Alarmi/Sources/Global/Extensions/UIViewController+.swift
@@ -1,0 +1,18 @@
+//
+//  UIViewController+.swift
+//  Alarmi
+//
+//  Created by Woody on 2022/07/15.
+//  Copyright © 2022 MoTe. All rights reserved.
+//
+
+import UIKit
+
+extension UIViewController {
+    
+    func wrappedByNavigationController() -> UINavigationController {
+        let navigationController = UINavigationController(rootViewController: self)
+        navigationController.navigationBar.prefersLargeTitles = true
+        return navigationController
+    }
+}
