+++ Alarmi/Alarmi/Sources/ViewControllers/TodayViewController.swift
@@ -12,6 +12,7 @@ class TodayViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        
         attribute()
     }
